requirements.txt
flask==2.3.3
gunicorn==21.2.0
pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.2
werkzeug==2.3.7